update outlook.vb
Replaced the Set objRegEx = CreateObject("VBScript.RegExp") line with the Set regex = CreateObject("VBScript.RegExp") line. This will make the variable names consistent, as regex is used throughout the rest of the code.

Removed the Global and MultiLine properties from the regular expression object, as they are not needed in this case.

Removed the If Not IsNumeric(strRoomNumber) Then ... End If block and replace it with a check using the regex.test method. This will allow you to handle cases where the room number is not found in the subject line, as well as cases where it is found but is not numeric.